requirements.txt
alembic==1.0.9
click==6.7
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.0
psycopg2==2.8.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.14.1
WTForms==2.2.1
wheel==0.33.1
